Create CardClassPyGameMerged
You have some options...
I've left comments throughout regarding which Rank you wish to use.;  Read comments in the document thoroughly before making deciding on which version you understand better, one is a lot of files, but simple code, the other is a great deal of typing that needs to  be streamlined.  NOTE::  If you did Pygame in class, you should do this one now in the version of your choosing.